Create ActivityTaskProgressModel which shows cummulative progress
- Move relevant code to common-lib instead of common-gui
