+++ common-gui/src/main/java/org/drugis/common/gui/task/TaskProgressBar.java
@@ -1,14 +1,15 @@
 package org.drugis.common.gui.task;
 
-import org.drugis.common.gui.TextProgressBar;
 import org.drugis.common.threading.Task;
+import org.drugis.common.threading.status.AbstractProgressModel;
+import org.drugis.common.threading.status.TaskProgressModel;
 
 @SuppressWarnings("serial")
 public class TaskProgressBar extends TextProgressBar {
-	public TaskProgressBar(TaskProgressModel model) {
+	public TaskProgressBar(AbstractProgressModel model) {
 		super(model);
 	}
-	
+
 	public TaskProgressBar(Task task) {
 		super(new TaskProgressModel(task));
 	}
