+++ common-lib/src/test/java/org/drugis/common/threading/status/TaskProgressModelTest.java
@@ -1,4 +1,4 @@
-package org.drugis.common.gui.task;
+package org.drugis.common.threading.status;
 
 import static org.easymock.EasyMock.verify;
 import static org.junit.Assert.assertEquals;
@@ -12,7 +12,6 @@ import java.util.Arrays;
 import java.util.List;
 
 import org.drugis.common.JUnitUtil;
-import org.drugis.common.gui.TextProgressModel;
 import org.drugis.common.threading.AbstractIterativeComputation;
 import org.drugis.common.threading.CompositeTask;
 import org.drugis.common.threading.FailureException;
@@ -30,7 +29,7 @@ public class TaskProgressModelTest {
 		public ListenerManager d_mgr = new ListenerManager(this);
 		private boolean d_finished = false;
 		private String d_str = null;
-		
+
 		public MockComposite() { }
 		public MockComposite(String str) { d_str = str; }
 		@Override public String toString() { return d_str != null ? d_str : mkString(); }
@@ -49,7 +48,7 @@ public class TaskProgressModelTest {
 		public boolean isAborted() { return false; }
 		public Throwable getFailureCause() { return null; }
 		public List<SimpleTask> getNextTasks() { return null; }
-		
+
 		public String mkString() {
 			return super.toString().substring(super.toString().indexOf('@'));
 		}
@@ -60,13 +59,13 @@ public class TaskProgressModelTest {
 			super(max);
 		}
 		public void doStep() {}
-		
+
 		@Override
 		public String toString() {
 			return super.toString().substring(super.toString().indexOf('@'));
 		}
 	}
-	
+
 	@Test
 	public void testGetSetTask() {
 		ShortComputation comp = new ShortComputation(100);
@@ -75,16 +74,16 @@ public class TaskProgressModelTest {
 		mod.setTask(task);
 		assertTrue(mod.getTask()== task);
 	}
-	
+
 	@Test
 	public void testSimpleProgress() {
 		IterativeTask task = new IterativeTask(new ShortComputation(10000));
 		task.setReportingInterval(10);
-		
+
 		TextProgressModel model = new TaskProgressModel(task);
 		assertEquals(task.toString() + " (waiting)", model.getText());
 		assertFalse(model.getDeterminate());
-		
+
 		task.run();
 		assertEquals(TaskProgressModel.DONE_TEXT, model.getText());
 		assertTrue(model.getDeterminate());
@@ -95,7 +94,7 @@ public class TaskProgressModelTest {
 	public void testSimpleProgressEvents() {
 		final IterativeTask task = new IterativeTask(new ShortComputation(10000));
 		task.setReportingInterval(2000);
-		
+
 		final TextProgressModel model = new TaskProgressModel(task);
 		final List<Double> progressEvents = new ArrayList<Double>();
 		final List<String> textEvents = new ArrayList<String>();
@@ -109,11 +108,11 @@ public class TaskProgressModelTest {
 				}
 			}
 		});
-		
+
 		task.run();
 		Double[] expected = new Double[] {0.0, 0.0, 0.2, 0.4, 0.6, 0.8, 1.0, 1.0};
 		assertEquals(Arrays.asList(expected), progressEvents);
-		
+
 		String[] expectedStr = new String[] {
 			task.toString() + ": ?",
 			task.toString() + ": 0%",
@@ -130,37 +129,37 @@ public class TaskProgressModelTest {
 	@Test
 	public void testDeterminate() {
 		final MockTask task = new MockTask();
-		
+
 		final TextProgressModel model = new TaskProgressModel(task);
 		task.start();
-		
+
 		PropertyChangeListener mock = JUnitUtil.mockListener(model, TextProgressModel.PROPERTY_DETERMINATE, false, true);
 		model.addPropertyChangeListener(mock);
 		task.progress(0, 1000);
 		verify(mock);
 	}
-	
+
 	@Test
 	public void testFinishedDeterminate() {
 		final MockTask task = new MockTask();
-		
+
 		final TextProgressModel model = new TaskProgressModel(task);
 		task.start();
-		
+
 		PropertyChangeListener mock = JUnitUtil.mockListener(model, TextProgressModel.PROPERTY_DETERMINATE, false, true);
 		model.addPropertyChangeListener(mock);
 		task.finish();
 		verify(mock);
 	}
-	
+
 	@Test
 	public void testFailed() {
 		final MockTask task = new MockTask();
-		
+
 		final TextProgressModel model = new TaskProgressModel(task);
 		task.start();
-		
-		String failText = task + " " + TaskProgressModel.FAILED_TEXT +": " + "Argh!"; 
+
+		String failText = task + " " + TaskProgressModel.FAILED_TEXT +": " + "Argh!";
 		PropertyChangeListener mock1 = JUnitUtil.mockListener(model, TextProgressModel.PROPERTY_DETERMINATE, false, true);
 		PropertyChangeListener mock2 = JUnitUtil.mockListener(model, TextProgressModel.PROPERTY_TEXT, null,	failText);
 		PropertyChangeListener mock3 = JUnitUtil.mockListener(model, TextProgressModel.PROPERTY_PROGRESS, 0.0, 0.0);
@@ -172,16 +171,16 @@ public class TaskProgressModelTest {
 		verify(mock2);
 		verify(mock3);
 	}
-	
+
 	@Test
 	public void testAborted() {
 		final MockTask task = new MockTask();
-		
+
 		final TextProgressModel model = new TaskProgressModel(task);
 		task.start();
-		
+
 		PropertyChangeListener mock1 = JUnitUtil.mockListener(model, TextProgressModel.PROPERTY_DETERMINATE, false, true);
-		PropertyChangeListener mock2 = JUnitUtil.mockListener(model, TextProgressModel.PROPERTY_TEXT, null, 
+		PropertyChangeListener mock2 = JUnitUtil.mockListener(model, TextProgressModel.PROPERTY_TEXT, null,
 				task.toString() + ": " + TaskProgressModel.ABORTED_TEXT);
 		PropertyChangeListener mock3 = JUnitUtil.mockListener(model, TextProgressModel.PROPERTY_PROGRESS, 0.0, 0.0);
 		model.addPropertyChangeListener(mock1);
@@ -192,16 +191,16 @@ public class TaskProgressModelTest {
 		verify(mock2);
 		verify(mock3);
 	}
-	
+
 	@Test
 	public void testPhasedProgress() {
 		MockComposite ctask = new MockComposite();
 		IterativeTask phase1 = new IterativeTask(new ShortComputation(10000));
 		IterativeTask phase2 = new IterativeTask(new ShortComputation(10000));
 		phase1.setReportingInterval(3000);
-		
+
 		final TextProgressModel model = new TaskProgressModel(ctask);
-		
+
 		assertEquals(ctask.toString() + " (waiting)", model.getText());
 		ctask.start();
 		assertEquals(ctask.toString() + ": ?", model.getText());
@@ -213,7 +212,7 @@ public class TaskProgressModelTest {
 				}
 			}
 		});
-		
+
 		ctask.d_mgr.firePhaseStarted(phase1);
 		phase1.run();
 		String title1 = ctask.toString() + " (" + phase1.toString() + ": ";
@@ -231,7 +230,7 @@ public class TaskProgressModelTest {
 				TaskProgressModel.DONE_TEXT
 		};
 		ctask.d_mgr.firePhaseFinished(phase1);
-		
+
 		ctask.d_mgr.firePhaseStarted(phase2);
 		assertEquals(false, model.getDeterminate());
 		ctask.d_mgr.firePhaseFinished(phase2);
@@ -245,7 +244,7 @@ public class TaskProgressModelTest {
 		MockComposite ctask = new MockComposite();
 		MockComposite phase1 = new MockComposite("phase1");
 		MockComposite phase2 = new MockComposite("phase2");
-		
+
 		final TextProgressModel model = new TaskProgressModel(ctask);
 		final List<String> text = new ArrayList<String>();
 		final List<Boolean> determinate = new ArrayList<Boolean>();
@@ -260,23 +259,23 @@ public class TaskProgressModelTest {
 
 		final List<String> expStrings = new ArrayList<String>();
 		final List<Boolean> expDeterminate = new ArrayList<Boolean>();
-		
+
 		ctask.start();
 		expStrings.add(ctask.toString() + ": ?");
 		expDeterminate.add(false);
-		
+
 		ctask.d_mgr.firePhaseStarted(phase1);
 		expStrings.add(ctask.toString() + " (" + phase1.toString() + ": ?)");
 		expDeterminate.add(false);
-		
+
 		phase1.d_mgr.fireTaskProgress(30, 100);
 		expStrings.add(ctask.toString() + " (" + phase1.toString() + ": 30%)");
 		expDeterminate.add(true);
-		
+
 		ctask.d_mgr.firePhaseStarted(phase2);
 		expStrings.add(ctask.toString() + " (" + phase1.toString() + ": 30%, " + phase2.toString() + ": ?)");
 		expDeterminate.add(false);
-		
+
 		assertEquals(expStrings, text);
 		assertEquals(expDeterminate, determinate);
 	}
